requirements.txt
# RelCheck — Python dependencies
# Install with: pip install -r requirements.txt
# In Colab: !pip install -r requirements.txt

# NLP
spacy>=3.7.0
# After install: python -m spacy download en_core_web_sm

# Vision-Language Models
transformers>=4.37.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
accelerate>=0.25.0
bitsandbytes>=0.41.0   # for 8-bit model loading on GPU

# LLM API (Mistral correction)
together>=1.2.0

# Metrics
python-Levenshtein>=0.21.0
nltk>=3.8.0

# Text processing
json-repair>=0.30.0        # robust LLM JSON recovery
pysbd>=0.3.4               # sentence boundary detection (EMNLP 2020)
rapidfuzz>=3.0.0           # fuzzy string matching

# Retry / resilience
tenacity>=8.2.0

# Data / Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
